split runner file: game scene
