+++ CMakeLists.txt
@@ -117,6 +117,7 @@ set(SOURCES
 	sources/System/IntroAnimation/IntroAnimationSystem.cpp
 	sources/Runner/SplashScreenScene.cpp
 	sources/Runner/TitleScreenScene.cpp
+	sources/Runner/GameScene.cpp
 	sources/Runner/CreditScene.cpp
 )
 add_executable(bomberman
